+++ src/layout/arrayAlign/types.ts
@@ -1,0 +1,4 @@
+import { AlignObject } from "../Align/Align";
+
+export type ILayoutTarget = AlignObject;
+export type ArrayAlignCallback = <T extends ILayoutTarget>(objects: T[], ...args: number[]) => T[]