+++ src/layout/types.ts
@@ -1,0 +1,3 @@
+export interface IBoundable {
+  getBounds(): Phaser.Geom.Rectangle;
+}