+++ changelog.md
@@ -1,3 +1,12 @@
 
+# 1.1.0
+- ViewPort component 
+- typeGuards isScene, isRectangle, hasGetBounds, hasWidthHeight
+- Layout for area layout
+- AlignMethod as command collectios for Align.applyMethod(alignMethod, obj, ...args)
+- align chains for array layouts - rows, columns, chains
+- ArrayAlignObject class for array layout without containers
+
+
 # 1.0.34
 Add `Align.anchorArea` method and bump version to 1.0.34